Rename the docker, docker_windows, and root_filesystem modules
