+++ scanpipe/pipelines/analyze_docker_windows.py
@@ -20,7 +20,7 @@
 # ScanCode.io is a free software code scanning tool from nexB Inc. and others.
 # Visit https://github.com/aboutcode-org/scancode.io for support and download.
 
-from scanpipe.pipelines.docker import Docker
+from scanpipe.pipelines.analyze_docker import Docker
 from scanpipe.pipes import docker
 from scanpipe.pipes import rootfs
 from scanpipe.pipes import windows
