+++ setup.cfg
@@ -136,9 +136,9 @@ console_scripts =
     run = scancodeio:combined_run
 
 scancodeio_pipelines =
-    analyze_docker_image = scanpipe.pipelines.docker:Docker
-    analyze_root_filesystem_or_vm_image = scanpipe.pipelines.root_filesystem:RootFS
-    analyze_windows_docker_image = scanpipe.pipelines.docker_windows:DockerWindows
+    analyze_docker_image = scanpipe.pipelines.analyze_docker:Docker
+    analyze_root_filesystem_or_vm_image = scanpipe.pipelines.analyze_root_filesystem:RootFS
+    analyze_windows_docker_image = scanpipe.pipelines.analyze_docker_windows:DockerWindows
     collect_strings_gettext = scanpipe.pipelines.collect_strings_gettext:CollectStringsGettext
     collect_symbols_ctags = scanpipe.pipelines.collect_symbols_ctags:CollectSymbolsCtags
     collect_symbols_pygments = scanpipe.pipelines.collect_symbols_pygments:CollectSymbolsPygments
